Align border with end of stripes
