+++ src/pages/learn/index.mdx
@@ -5,8 +5,9 @@ title: Learn
 import { Button } from '@/app/conf/_design-system/button';
 
 import { NavbarFixed } from '../../components/navbar/navbar-fixed'
-import { TocHero, TocHeroContents } from '../../components/toc-hero';
-import { TeaserSection, LearnHeroStripes } from '../../components/learn-aggregator'
+import { TocHero, TocHeroContents } from '../../components/toc-hero'
+import { TeaserSection } from '../../components/learn-aggregator/teaser-section'
+import { LearnHeroStripes } from '../../components/learn-aggregator/learn-hero-stripes'
 import { pagesBySection } from '../../components/learn-aggregator/learn-pages'
 import { CommonQuestionsSection } from '../../components/learn-aggregator/common-questions'
 import { TrainingCoursesSection } from '../../components/learn-aggregator/training-courses'
